Separate out testnet/sandbox create account into sep files
